requirements.txt
Django==3.2.3
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-camel-case
flake8==3.7.9
google-auth==1.33.0
gunicorn==20.0.4
Markdown==3.1.1
psycopg2==2.8.6
ptvsd==4.3.2
pytest==5.4.1
pytest-django==3.9.0
pytest-factoryboy==2.0.3